compile

+++ src/edu/washington/escience/myria/perfenforce/encoding/ScalingAlgorithmEncoding.java
@@ -17,6 +17,6 @@ public class ScalingAlgorithmEncoding {
   public double beta;
   public double kp;
   public double ki;
-  public double w;
+  public int w;
   public double lr;
 }
